add example of how to add posthog api key
