+++ packages/server/package.json
@@ -60,6 +60,7 @@
     "passport-google-oauth20": "^2.0.0",
     "pg": "^8.5.1",
     "pg-query-stream": "^3.4.2",
+    "posthog-node": "^1.1.5",
     "prom-client": "^13.1.0",
     "redis": "^3.1.1",
     "sanitize-html": "^2.4.0",
